Rename files to match new class names
