+++ src/utils/index.ts
@@ -11,6 +11,6 @@ export * from './setters';
 export * from './sound';
 export * from './state';
 export * from './translate';
-export * from './vibrator';
+export * from './toyclient';
 export * from './wait';
 export * from './toyactuator';
